refactor: introduce 1 more layer of indirection to link modal
1. previously, it was taking editor directly which made it harder to
reuse
2. now, we want to show the modal also on json forms so we need to
abstract the `editor` prop
