+++ apps/studio/src/components/PageEditor/MenuBar/AccordionMenuBar.tsx
@@ -26,9 +26,9 @@ import {
   IconMergeCells,
   IconSplitCell,
 } from "~/components/icons"
-import { LinkEditorModal } from "../LinkEditorModal"
 import { TableSettingsModal } from "../TableSettingsModal"
 import { MenuBar } from "./MenuBar"
+import { TiptapLinkEditorModal } from "./TiptapLinkEditorModal"
 
 export const AccordionMenuBar = ({ editor }: { editor: Editor }) => {
   const {
@@ -215,7 +215,7 @@ export const AccordionMenuBar = ({ editor }: { editor: Editor }) => {
         onClose={onTableSettingsModalClose}
       />
 
-      <LinkEditorModal
+      <TiptapLinkEditorModal
         editor={editor}
         isOpen={isLinkModalOpen}
         onClose={onLinkModalClose}
